+++ fB_web/types/shims.d.ts
@@ -1,0 +1,5 @@
+declare module 'mssql' {
+  const sql: any
+  export = sql
+}
+